% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{using_account}
\alias{using_account}
\title{Determine whether a given user account is being used}
\usage{
using_account(username)
}
\arguments{
\item{username}{Username to check for}
}
\value{
A logical value
}
\description{
Determine whether a given user account is being used
}
\examples{
using_account("root")
}
