% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor.R
\name{on_bioconductor}
\alias{on_bioconductor}
\title{Detect Bioconductor}
\usage{
on_bioconductor()
}
\value{
A logical value
}
\description{
Detect Bioconductor
}
\examples{
on_bioconductor()
}
\references{
Check the Bioconductor Build System: https://github.com/Bioconductor/BBS/
}
