% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda.R
\name{using_conda}
\alias{using_conda}
\title{Determine whether Conda environment is being used}
\usage{
using_conda(env = NULL)
}
\arguments{
\item{env}{Optional environment name to compare against}
}
\value{
A logical value
}
\description{
Determine whether Conda environment is being used
}
\examples{
# Check if Conda is being used (regardless of environment name)
using_conda()

# Check if the 'dev' Conda environment is being used
using_conda(env = "dev")
}
