% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{using_container}
\alias{using_container}
\alias{using_docker_container}
\alias{using_podman_container}
\alias{using_kubernetes}
\alias{using_runpod}
\title{Detect container environments}
\usage{
using_container()

using_docker_container()

using_podman_container()

using_kubernetes()

using_runpod()
}
\value{
A logical value
}
\description{
Detect container environments
}
\examples{
using_container()
using_docker_container()
using_podman_container()
using_kubernetes()
using_runpod()
}
