% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpu.R
\name{using_x86_cpu}
\alias{using_x86_cpu}
\alias{using_arm_cpu}
\title{Processor Checks}
\usage{
using_x86_cpu()

using_arm_cpu()
}
\value{
A logical value

A logical value
}
\description{
\code{using_x86_cpu()} checks whether the machine uses an x86
processor or not

\code{using_arm_cpu()} checks whether the machine uses an ARM-based
processor or not
}
\examples{
using_x86_cpu()
using_arm_cpu()
}
