% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.R
\name{on_cran}
\alias{on_cran}
\title{Detect CRAN}
\usage{
on_cran(CRAN_pattern = "^_R_", n_CRAN_envvars = 5L)
}
\value{
A logical value
}
\description{
Detect CRAN
}
\examples{
on_cran()
Sys.setenv("_R_1" = 1)
Sys.setenv("_R_2" = 2)
on_cran(n_CRAN_envvars = 2L) # TRUE
}
