% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databricks.R
\name{using_databricks}
\alias{using_databricks}
\title{Detect Databricks Runtime Environment}
\usage{
using_databricks()
}
\value{
A logical value
}
\description{
Detect Databricks Runtime Environment
}
\examples{
using_databricks()
}
