% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{using_github_codespaces}
\alias{using_github_codespaces}
\alias{using_github_actions}
\title{GitHub Environments}
\usage{
using_github_codespaces(name = NULL)

using_github_actions()
}
\arguments{
\item{name}{Optional codespace name to compare against}
}
\value{
A logical value
}
\description{
\code{using_github_codespaces} checks whether code is being run in a
\href{https://github.com/features/codespaces}{GitHub Codespaces} environment.

\code{using_github_actions()} reports whether GitHub Actions is being
used
}
\examples{
# Check if GitHub Codespaces is being used
using_github_codespaces()

# Check if the 'octocat-literate-space-parakeet-mld5' codespace is being used
using_python_venv(env = "octocat-literate-space-parakeet-mld5")
using_github_actions()
}
