% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online.R
\name{is_online}
\alias{is_online}
\title{Check whether the machine is connected to the internet}
\usage{
is_online()
}
\value{
A logical value
}
\description{
Check whether the machine is connected to the internet
}
\examples{
is_online()
}
