% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networking.R
\name{online}
\alias{online}
\alias{using_host}
\title{Checks related to networking}
\usage{
online()

using_host(hostname)
}
\arguments{
\item{hostname}{String containing a hostname or hostnames to check}
}
\value{
A logical value
}
\description{
\code{online()} uses \link[curl:nslookup]{curl::has_internet} to check whether the
machine is connected to the internet
}
\examples{
online()
using_host("somehost.fakedomain.com")
}
