% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nix.R
\name{using_nix_shell}
\alias{using_nix_shell}
\title{Detect \code{Nix} Shell}
\usage{
using_nix_shell(pure = NULL)
}
\arguments{
\item{pure}{Whether or not the environment is pure, meaning most environment
variables have been cleared before the shell started.}
}
\value{
A logical value
}
\description{
\code{using_nix_shell()} checks whether code is running within an
environment defined by a \href{https://nixos.org/manual/nix/unstable/command-ref/nix-shell.html}{Nix expression}.
}
\examples{
# Check for Nix
using_nix_shell()

# Check for Nix in a pure environment
using_nix_shell(pure = TRUE)
}
