% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online.R
\name{online}
\alias{online}
\title{Check whether the machine is connected to the internet}
\usage{
online()
}
\value{
A logical value
}
\description{
Check whether the machine is connected to the internet
}
\examples{
online()
}
