% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option.R
\name{using_option}
\alias{using_option}
\title{Determine whether a global option is being used}
\usage{
using_option(x, value = NULL)
}
\arguments{
\item{x}{Option name}

\item{value}{Optional value to compare against}
}
\value{
A logical value
}
\description{
Determine whether a global option is being used
}
\examples{
using_option("width")

using_option("boot.parallel", value = "multicore")
}
