% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{using_os}
\alias{using_os}
\alias{using_linux}
\alias{using_macos}
\alias{using_solaris}
\alias{using_windows}
\title{Tests for operating systems}
\usage{
using_os(os = c("linux", "macos", "solaris", "windows"))

using_linux()

using_macos()

using_solaris()

using_windows()
}
\arguments{
\item{os}{Operating system name. One of \code{"linux"}, \code{"macos"}, \code{"solaris"}, or
\code{"windows"}}
}
\value{
A logical value
}
\description{
Tests for operating systems
}
\examples{
using_os(os = "linux")
using_linux()
using_macos()
using_solaris()
using_windows()
}
