% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positron.R
\name{using_positron}
\alias{using_positron}
\alias{using_positron_desktop}
\alias{using_positron_server}
\title{Positron environments}
\usage{
using_positron(mode = "any")

using_positron_desktop()

using_positron_server()
}
\arguments{
\item{mode}{Optional argument specifying whether Positron is being used in
\code{"desktop"} mode or in \code{"server"} mode.}
}
\value{
A logical value
}
\description{
These functions enable you to determine whether code is being run in the
presence of various features of the
\href{https://positron.posit.co}{Positron IDE}

\code{using_positron()} determines whether code is being run in Positron.
\code{using_positron_desktop()}, \code{using_positron_server()} are helpers to
determine whether those specific
environments are being used.
}
\examples{
using_rstudio()
}
