% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_venv.R
\name{using_python_venv}
\alias{using_python_venv}
\title{Determine whether a Python virtual environment is being used}
\usage{
using_python_venv(env = NULL)
}
\arguments{
\item{env}{Optional environment name to compare against}
}
\value{
A logical value
}
\description{
Determine whether a Python virtual environment is being used
}
\examples{
# Check if a Python virtual environment is being used
using_python_venv()

# Check if the 'dev' virtual environment is being used
using_python_venv(env = "dev")
}
