% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{using_quarto}
\alias{using_quarto}
\title{Quarto documents}
\usage{
using_quarto()
}
\value{
A logical value
}
\description{
\code{using_quarto()} determines whether code is being run within a Quarto
document
}
\note{
The \code{is_using_quarto()} function in the \href{https://cran.r-project.org/package=quarto}{quarto}
package can be used to determine whether your are in a quarto project.
}
\examples{
using_quarto()
}
