% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{using_interactive_session}
\alias{using_interactive_session}
\title{R session information}
\usage{
using_interactive_session()
}
\description{
\code{using_interactive_session()} determines whether or not R is
being run interactively. Under the hood, \link[rlang:is_interactive]{rlang::is_interactive} is used.
}
\examples{
using_interactive_session()
}
