% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{using_rstudio}
\alias{using_rstudio}
\alias{using_rstudio_jobs}
\title{RStudio environments}
\usage{
using_rstudio()

using_rstudio_jobs()
}
\value{
A logical value
}
\description{
RStudio environments
}
\examples{
using_rstudio()
using_rstudio_jobs()
}
