% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitch-package.R
\docType{package}
\name{sitch-package}
\alias{sitch}
\alias{sitch-package}
\title{sitch: Checks for Various Computing Environments}
\description{
sitch is a collection of lightweight functions that can be used to determine the computing environment in which your code is running. This includes Continuous Integration (CI) environments, containers, different operating systems, and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/briandconnelly/sitch}
  \item \url{https://briandconnelly.github.io/sitch/}
  \item Report bugs at \url{https://github.com/briandconnelly/sitch/issues}
}

}
\author{
\strong{Maintainer}: Brian Connelly \email{bdc@bconnelly.net} (\href{https://orcid.org/0000-0002-9948-0379}{ORCID}) [copyright holder]

}
\keyword{internal}
