% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envvar.R
\name{using_envvar}
\alias{using_envvar}
\title{Determine whether an environment variable is being used}
\usage{
using_envvar(x, value = NULL)
}
\arguments{
\item{x}{Environment variable}

\item{value}{Optional value to compare against}
}
\value{
A logical value
}
\description{
Determine whether an environment variable is being used
}
\examples{
using_envvar("NOT_CRAN")

using_envvar("CI", "true")
}
