% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{using_windows}
\alias{using_windows}
\title{Tests for 'Windows' environments}
\usage{
using_windows()
}
\value{
A logical value
}
\description{
Tests for 'Windows' environments
}
\examples{
using_windows()
}
