% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ami-package.R
\docType{package}
\name{ami-package}
\alias{ami}
\alias{ami-package}
\title{ami: Checks for Various Computing Environments}
\description{
A collection of lightweight functions that can be used to determine the computing environment in which your code is running. This includes operating systems, continuous integration (CI) environments, containers, and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/briandconnelly/ami}
  \item \url{https://briandconnelly.github.io/ami/}
  \item Report bugs at \url{https://github.com/briandconnelly/ami/issues}
}

}
\author{
\strong{Maintainer}: Brian Connelly \email{bdc@bconnelly.net} (\href{https://orcid.org/0000-0002-9948-0379}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID}) [contributor]
  \item Lluís Revilla Sancho \email{lluis.revilla@gmail.com} (\href{https://orcid.org/0000-0001-9747-2570}{ORCID}) [contributor]
}

}
\keyword{internal}
