% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{using_container}
\alias{using_container}
\alias{using_docker_container}
\title{Detect container environments}
\usage{
using_container()

using_docker_container()
}
\value{
A logical value
}
\description{
Detect container environments
}
\examples{
using_container()
using_docker_container()
}
