% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpu.R
\name{using_cpu}
\alias{using_cpu}
\alias{using_x86_cpu}
\alias{using_arm_cpu}
\title{Processor Checks}
\usage{
using_cpu(arch = c("arm", "x86"))

using_x86_cpu()

using_arm_cpu()
}
\arguments{
\item{arch}{CPU architecture name. Either \code{"arm"} or \code{"x86"}.}
}
\value{
A logical value
}
\description{
\code{using_cpu()} checks whether the machine uses the given type
of processor or not.

\code{using_x86_cpu()} checks whether the machine uses an x86
processor

\code{using_arm_cpu()} checks whether the machine uses an ARM-based
processor
}
\examples{
using_arm_cpu()
using_x86_cpu()
using_arm_cpu()
}
