% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{using_github_actions}
\alias{using_github_actions}
\title{Detect GitHub Workflow environments}
\usage{
using_github_actions()
}
\value{
A logical value
}
\description{
Detect GitHub Workflow environments
}
\examples{
using_github_actions()
}
