% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vscode.R
\name{using_vscode}
\alias{using_vscode}
\title{Detect whether code is running in Visual Studio Code}
\usage{
using_vscode()
}
\value{
A logical value
}
\description{
Detect whether code is running in Visual Studio Code
}
\examples{
using_vscode()
}
